+++ module/webroot/file_selector.css
@@ -1,18 +1,3 @@
-.file-selector-modal {
-    display: none;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0, 0, 0, 0.5);
-    transition: opacity 0.3s ease;
-    opacity: 0;
-    z-index: 2000;
-    justify-content: center;
-    align-items: center;
-}
-
 .file-selector {
     width: 90%;
     max-width: 600px;
