+++ module/webroot/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>System App Nuker</title>
-    <script type="module" src="script.js" defer></script>
+    <script type="module" src="script.js"></script>
     <link rel="stylesheet" type="text/css" href="/mmrl/insets.css" />
     <link rel="stylesheet" type="text/css" href="styles.css">
     <link rel="stylesheet" type="text/css" href="file_selector.css">
@@ -148,7 +148,7 @@
     </div>
 
     <!-- File Selector Modal -->
-    <div class="file-selector-modal">
+    <div class="file-selector-modal modal">
         <div class="file-selector">
             <div class="file-selector-header">
                 <button class="back-button">
